% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_yaml_fragments.R
\name{extract_yaml_fragments}
\alias{extract_yaml_fragments}
\title{Extract all YAML fragments from a file}
\usage{
extract_yaml_fragments(file, text, delimiterRegEx = "^---$",
  ignoreOddDelimiters = FALSE, encoding = "UTF-8", silent = TRUE)
}
\arguments{
\item{file}{The path to a file to scan; if provided, takes precedence
over \code{text}.}

\item{text}{A character vector to scan, where every element should
represent one line in the file; can be specified instead of \code{file}.}

\item{delimiterRegEx}{The regular expression used to locate YAML
fragments.}

\item{ignoreOddDelimiters}{Whether to throw an error (FALSE) or
delete the last delimiter (TRUE) if an odd number of delimiters is
encountered.}

\item{encoding}{The encoding to use when calling \code{\link[=readLines]{readLines()}}. Set to
NULL to let \code{\link[=readLines]{readLines()}} guess.}

\item{silent}{Whether to be silent (TRUE) or informative (FALSE).}
}
\value{
A list of character vectors.
}
\description{
These function extracts all YAML fragments from a file,
returning a list of character vectors containing the extracted
fragments.
}
\examples{
yum::extract_yaml_fragments((text=c("---", "First: YAML fragment", "---",
                                    "Outside of YAML",
                                    "---", "Second: YAML fragment", "---",
                                    "Also outside of YAML"));

}
