% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_list_of_lists.R
\name{flatten_list_of_lists}
\alias{flatten_list_of_lists}
\title{Flatten a list of lists to a list of atomic vectors}
\usage{
flatten_list_of_lists(x)
}
\arguments{
\item{x}{The list of lists.}
}
\value{
A list of atomic vectors.
}
\description{
This function takes a hierarchical structure of lists and
extracts all atomic vectors, returning one flat list of all
those vectors.
}
\examples{
### First create a list of lists
listOfLists <-
  list(list(list(1:3, 8:5), 7:7), list(1:4, 8:2));
yum::flatten_list_of_lists(listOfLists);
}
